Removes is_exhausted() and reference iterators
is_exhausted() is not needed and the implementation
of iterators is easier if we only have to implement next().

The STL iterators that produced references caused trouble
if they we're concat-ed with iterators producing values.
This could cause the value producing iterators to
discard their values and refer to garbage memory.
All iterators now produce values which in some cases
cause an superfluent copy, but the peace of mind
when using the library greatly outweighes this cost.